Fix travis CI build for S4
With gradle 2.1, running the `gradle wrapper` tool in the S4 base directory does not work
due to a missing parameter in the s4 build.gradle file.
Travis CI runs gradle 2.1 by default from september 2014. As a result, the S4 build fails.

This patch is a workaround that includes the gradle 1.4 dependencies in the samoa repo and
copies them appropriately during the travis CI build.