Fix end of stream management
When reaching the end of the stream, we mark the last event accordingly and
set the finished status on the processor. Otherwise the last event is never marked as such,
and tests that wait for this particular event don't properly terminate.

Updated test expectations to match new behaviour (no event is sent after stream is consumed).